Abstract combine_expressions in utils, and make npm, maven and rubygems use the shared utils function, force the test to set license_epxression of rubygems #1480
