+++ tests/packagedcode/data/rubygems/gem/action_tracker-1.0.2.gem.json
@@ -37,7 +37,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "mit", 
     "declared_license": [
       "MIT"
     ],
