+++ tests/packagedcode/data/rubygems/gem/m2r-2.1.0.gem.json
@@ -51,7 +51,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "mit", 
     "declared_license": [
       "MIT"
     ],
