+++ tests/packagedcode/data/rubygems/gem/new/dependabot-omnibus-0.87.15.gem.json
@@ -30,7 +30,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "unknown", 
     "declared_license": [
       "Nonstandard"
     ],
