+++ tests/packagedcode/data/rubygems/gem/new/hiredis-0.6.3-java.gem.json
@@ -30,7 +30,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "bsd-new", 
     "declared_license": [
       "BSD-3-Clause"
     ],
