+++ tests/packagedcode/data/rubygems/gem/new/int_time-0.0.2.gem.json
@@ -30,7 +30,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "mit", 
     "declared_license": [
       "MIT"
     ],
