+++ tests/packagedcode/data/rubygems/gem/new/jaro_winkler-1.5.1-java.gem.json
@@ -30,7 +30,7 @@
     "code_view_url": null, 
     "vcs_url": null, 
     "copyright": null, 
-    "license_expression": null, 
+    "license_expression": "mit", 
     "declared_license": [
       "MIT"
     ],
